+++ src/content/model/TaskEvent.ts
@@ -6,6 +6,7 @@ export class TaskEvent {
     asin: string;
     avg_rating: number;
     delivery_info: string;
+    out_of_stock: string;
     n_ratings: number;
     name: string;
     position: number;
@@ -51,6 +52,7 @@ export class TaskEvent {
                 this.asin = item.asin;
                 this.avg_rating = item.avgRating;
                 this.delivery_info = item.deliveryInfo;
+                this.out_of_stock = item.outOfStockTxt;
                 this.n_ratings = item.nReviews;
                 this.name = item.name;
                 this.position = item.position;